+++ contracts/interfaces/structs/PoolsharkStructs.sol
@@ -55,7 +55,7 @@ interface PoolsharkStructs {
         address tokenOut;
         uint160 startPrice;
         uint16  swapFee;
-        uint8   poolTypeId;
+        uint16  poolTypeId;
     }
 
     struct MintLimitParams {
